+++ Papr/Commons/Papr.swift
@@ -1,0 +1,12 @@
+//
+//  Papr.swift
+//  Papr
+//
+//  Created by Joan Disho on 22.01.18.
+//  Copyright © 2018 Joan Disho. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+enum Papr {}
